+++ src/main/org/firebirdsql/util/FirebirdSupportInfo.java
@@ -562,6 +562,20 @@ public final class FirebirdSupportInfo {
     }
 
     /**
+     * @return {@code true} when this Firebird version supports NBackup backup with GUID
+     */
+    public boolean supportsNBackupWithGuid() {
+        return isVersionEqualOrAbove(4, 0);
+    }
+
+    /**
+     * @return {@code true} when this Firebird version supports NBackup in-place restore
+     */
+    public boolean supportsNBackupInPlaceRestore() {
+        return isVersionEqualOrAbove(4, 0);
+    }
+
+    /**
      * @param serverVersion
      *         Server version
      * @return FirebirdVersionSupport instance